Started adding renderer
